fix(#471): address CodeRabbit review - fallback metadata guard, scrobbleSong catch vars, extract fetchWithTimeout
