+++ src/main/other/lastFm/sendFavoritesDataToLastFM.ts
@@ -10,21 +10,11 @@ import logger from '../../logger';
 import { checkIfConnectedToInternet } from '../../main';
 import hashText from '../../utils/hashText';
 import getLastFmAuthData from './getLastFMAuthData';
+import { LASTFM_REQUEST_TIMEOUT_MS, fetchWithTimeout } from './lastFmUtils';
 
 type Method = 'track.love' | 'track.unlove';
 
-const LASTFM_REQUEST_TIMEOUT_MS = 10_000;
 const LASTFM_BASE_URL = 'https://ws.audioscrobbler.com/2.0/';
-
-const fetchWithTimeout = async (url: URL, init: RequestInit, timeoutMs: number) => {
-  const controller = new AbortController();
-  const timer = setTimeout(() => controller.abort(), timeoutMs);
-  try {
-    return await fetch(url, { ...init, signal: controller.signal });
-  } finally {
-    clearTimeout(timer);
-  }
-};
 
 const generateApiSignature = (method: Method, authData: AuthData, params: LoveParams) => {
   const { LAST_FM_API_KEY, LAST_FM_SHARED_SECRET, SESSION_KEY } = authData;
